backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
pillow
pydantic